refactor : CategoryContents component
Related to : #58
