+++ src/app/layout.tsx
@@ -4,8 +4,6 @@ import { Noto_Sans_KR } from 'next/font/google';
 import '@/styles/global.css';
 import Header from '@/components/layout/header/Header';
 import Footer from '@/components/layout/Footer';
-import RecentlyViewedBooks from '@/components/layout/RecentlyViewedBooks';
-// import styles from '@/styles/layout/layout.module.css';
 
 const inter = Noto_Sans_KR({ subsets: ['latin'] });
