+++ src/components/category/categoryContents/CategoryContents.tsx
@@ -2,7 +2,7 @@ import styles from '@/styles/category/categoryContents/CategoryContents.module.c
 import SortBar from './SortBar';
 import CategoryBookItemBox from './CategoryBookItemBox';
 import Pagination from '@/components/common/Pagination';
-import CategorySlider from './CategorySlider';
+import CategorySlider from './categorySlider/CategorySlider';
 
 export default function CategoryContents() {
 	return (
